pyproject.toml
[tool.poetry]
name = "taskflowai"
version = "0.5.4"
description = "TaskFlowAI is a lightweight python framework for building LLM based pipelines and multi-agent teams"
authors = ["Philippe Andre Page <philippeandrepage@gmail.com>"]
readme = "README.md"
packages = [{include = "taskflowai"}]
license = "Apache 2.0"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[tool.poetry.dependencies]
python = "^3.10,<3.13"
requests = "^2.5.0"
pydantic = "^2.0"
anthropic = "^0.31"
openai = "^1.0"
cohere = "^5.0"
beautifulsoup4 = "^4.12"
tqdm = "^4.65"
python-dotenv = "^1.0"
PyYAML = "^6.0"
ollama = "^0.1.0"
lxml = "^4.9"
halo = "^0.0.31"
groq = "^0.11.0"
numpy = ">=1,<2"
elevenlabs = "^1.0.0"
faiss-cpu = "^1.8.0"
pyyaml = "^6.0"
fredapi = "^0.5.0"
yfinance = "^0.2.20"
yahoofinance = "^0.0.2"
pinecone = "^5.3"
sentence_splitter = "^1.4"
igraph = "^0.11"
leidenalg = "^0.10"
fake-useragent = "^1.5"

[tool.poetry.extras]
langchain_tools = [
    "langchain-core",
    "langchain-community",
    "langchain-openai"
]
matplotlib_tools = [
    "matplotlib"
]
yahoo_finance_tools = [
    "yfinance",
    "yahoofinance",
    "pandas"
]
fred_tools = [
    "fredapi",
    "pandas"
]

[tool.poetry.dependencies.langchain-core]
version = "^0.3.0"
optional = true

[tool.poetry.dependencies.langchain-community]
version = "^0.3.0"
optional = true

[tool.poetry.dependencies.langchain-openai]
version = "^0.2.0"
optional = true

[tool.poetry.dependencies.matplotlib]
version = "^3.9"
optional = true

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0,<9.0.0"
black = ">=23.0,<25.0"
isort = ">=5.0,<6.0"
mypy = ">=1.0,<2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.urls]
"Homepage" = "https://github.com/philippe-page/taskflowai/"
"Bug Tracker" = "https://github.com/philippe-page/taskflowai/issues"
"Documentation" = "https://taskflowai.org"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]